backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
rdflib==7.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.8.5
cloudinary==1.41.0
